requirements.txt
streamlit
google-generativeai
langdetect 